refactor: pull out center browser
